Add ViewModel for total holdings
