+++ app/build.gradle
@@ -148,7 +148,7 @@ android {
       debuggable true
       versionNameSuffix "-DEBUG"
       ext.enableCrashlytics = false
-      minifyEnabled true
+      minifyEnabled false
       proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"
     }
   }
@@ -170,6 +170,13 @@ android {
     exclude "META-INF/LICENSE.txt"
     exclude "META-INF/NOTICE.txt"
   }
+    
+  lintOptions {
+        checkReleaseBuilds false
+        // Or, if you prefer, you can continue to check for errors in release builds,
+        // but continue the build even when errors are found:
+        abortOnError false
+    }
 }
 
 kapt {
@@ -204,7 +211,8 @@ dependencies {
   implementation "javax.annotation:javax.annotation-api:1.3.2"
 
   implementation "com.google.dagger:dagger:$DAGGER_VERSION"
-  kapt "com.google.dagger:dagger-compiler:$DAGGER_VERSION"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+    kapt "com.google.dagger:dagger-compiler:$DAGGER_VERSION"
 
   implementation ("com.squareup.okhttp3:okhttp:$OKHTTP_VERSION") {
     force = true
