+++ app/src/main/kotlin/com/github/premnirmal/ticker/home/HomeFragment.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.ViewGroup.MarginLayoutParams
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import com.github.premnirmal.ticker.base.BaseFragment
 import com.github.premnirmal.ticker.components.AsyncBus
@@ -48,24 +49,29 @@ class HomeFragment : BaseFragment(), ChildFragment, PortfolioFragment.Parent {
   private var attemptingFetch = false
   private var fetchCount = 0
   private lateinit var adapter: HomePagerAdapter
+  private lateinit var totalHoldingsViewModel: TotalHoldingsViewModel
 
   private val subtitleText: String
     get() = getString(R.string.last_and_next_fetch, stocksProvider.lastFetched(), stocksProvider.nextFetch())
 
-  private fun totalHoldings(): String
-  {
-    val (totalHolding, totalQuotesWithPosition) = stocksProvider.totalHoldings()
-    return getString(R.string.total_holdings, totalHolding, totalQuotesWithPosition)
+  private val totalHoldingsText: String
+    get() {
+      val (totalHolding, totalQuotesWithPosition) = totalHoldingsViewModel.getTotalHoldings()
+      return getString(R.string.total_holdings, totalHolding, totalQuotesWithPosition)
   }
 
-  private fun totalGainLoss(): Pair<String, String>
-  {
-    return stocksProvider.totalGainLoss()
+  private val totalGainLossText: Pair<String, String>
+    get() {
+      return totalHoldingsViewModel.getTotalGainLoss()
   }
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     Injector.appComponent.inject(this)
+
+    // Set up the ViewModel for the total holdings.
+    val factory = TotalHoldingsViewModelFactory(stocksProvider)
+    totalHoldingsViewModel = ViewModelProvider(this, factory).get(TotalHoldingsViewModel::class.java)
   }
 
   override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
@@ -83,8 +89,8 @@ class HomeFragment : BaseFragment(), ChildFragment, PortfolioFragment.Parent {
     view_pager.adapter = adapter
     tabs.setupWithViewPager(view_pager)
     subtitle.text = subtitleText
-    totalHoldings.text = totalHoldings()
-    val (totalGainStr, totalLossStr) = totalGainLoss()
+    totalHoldings.text = totalHoldingsText
+    val (totalGainStr, totalLossStr) = totalGainLossText
     totalGain.text = totalGainStr
     totalLoss.text = totalLossStr
   }
@@ -111,8 +117,8 @@ class HomeFragment : BaseFragment(), ChildFragment, PortfolioFragment.Parent {
     tabs.visibility = if (widgetDataProvider.hasWidget()) View.VISIBLE else View.INVISIBLE
     adapter.notifyDataSetChanged()
     subtitle.text = subtitleText
-    totalHoldings.text = totalHoldings()
-    val (totalGainStr, totalLossStr) = totalGainLoss()
+    totalHoldings.text = totalHoldingsText
+    val (totalGainStr, totalLossStr) = totalGainLossText
     totalGain.text = totalGainStr
     totalLoss.text = totalLossStr
   }
