+++ app/src/main/kotlin/com/github/premnirmal/ticker/model/IStocksProvider.kt
@@ -55,8 +55,4 @@ interface IStocksProvider {
     ticker: String,
     holding: Holding
   )
-
-  fun totalHoldings(): Pair<String, Int>
-
-  fun totalGainLoss(): Pair<String, String>
 }
