+++ gradle.properties
@@ -17,4 +17,4 @@ org.gradle.caching=true
 android.enableBuildCache=true
 android.useAndroidX=true
 android.enableJetifier=true
-android.enableUnitTestBinaryResources=true
+android.enableUnitTestBinaryResources=false
